+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.0.4
+* Add Claude 4 Sonnet
+* Add Claude 4 Opus
+* Deprecated Flag for Claude 3.7 Sonnet (use claude-4-sonnet instead)
+* Deprecated Flag for 4o, 4o-mini, o1, o1-pro, o1-mini, o3-mini (use o4-mini & o3 & 4.1 & 4.1-mini instead)
+* Migrated gemini-2.5-flash from xxx-4-17 to xxx-05-20
+
+# 1.0.3
+* Fixes
+
 # 1.0.2
 * Const connections & Inception